Modify style homepage
